+++ drivers/usb/core/hcd.c
@@ -3124,7 +3124,7 @@ int usb_hcd_setup_local_mem(struct usb_hcd *hcd, phys_addr_t phys_addr,
 	 * It's not backed by system memory and thus there's no kernel mapping
 	 * for it.
 	 */
-	err = gen_pool_add_virt(hcd->localmem_pool, (unsigned long)local_mem,
+	err = gen_pool_add_virt(hcd->localmem_pool, (uintptr_t)local_mem,
 				dma, size, dev_to_node(hcd->self.sysdev));
 	if (err < 0) {
 		dev_err(hcd->self.sysdev, "gen_pool_add_virt failed with %d\n",
